Move unit tests to avoid running it repeatedly
