+++ tests/self_join_tests.rs
@@ -87,3 +87,50 @@ pub async fn create_metadata(db: &DatabaseConnection) -> Result<(), DbErr> {
 
     Ok(())
 }
+
+mod tests {
+    use super::self_join::*;
+    use pretty_assertions::assert_eq;
+    use sea_orm::*;
+    use uuid::Uuid;
+
+    #[test]
+    fn find_linked_001() {
+        let self_join_model = Model {
+            uuid: Uuid::default(),
+            uuid_ref: None,
+            time: None,
+        };
+
+        assert_eq!(
+            self_join_model
+                .find_linked(SelfReferencingLink)
+                .build(DbBackend::MySql)
+                .to_string(),
+            [
+                r#"SELECT `self_join`.`uuid`, `self_join`.`uuid_ref`, `self_join`.`time`"#,
+                r#"FROM `self_join`"#,
+                r#"INNER JOIN `self_join` AS `r0` ON `r0`.`uuid_ref` = `self_join`.`uuid`"#,
+                r#"WHERE `r0`.`uuid` = '00000000-0000-0000-0000-000000000000'"#,
+            ]
+            .join(" ")
+        );
+    }
+
+    #[test]
+    fn find_also_linked_001() {
+        assert_eq!(
+            Entity::find()
+                .find_also_linked(SelfReferencingLink)
+                .build(DbBackend::MySql)
+                .to_string(),
+            [
+                r#"SELECT `self_join`.`uuid` AS `A_uuid`, `self_join`.`uuid_ref` AS `A_uuid_ref`, `self_join`.`time` AS `A_time`,"#,
+                r#"`r0`.`uuid` AS `B_uuid`, `r0`.`uuid_ref` AS `B_uuid_ref`, `r0`.`time` AS `B_time`"#,
+                r#"FROM `self_join`"#,
+                r#"LEFT JOIN `self_join` AS `r0` ON `self_join`.`uuid_ref` = `r0`.`uuid`"#,
+            ]
+            .join(" ")
+        );
+    }
+}
